Added backup and restore scripts for private stuff
They, in turn, backup and restore gpg, ssh, electrum, and anything in
the directory .private
Works properly only from the home folder, but they can backup securley
anywhere.
Encryption is done using GPG, so make sure you are backing up the key
you are using to encrypt, or you may be unable to get it back!